broke it, restarting...
failing to use git branches, working directly in master finally slew me. revuilt everything and re-implemented character movement in a more satisfactory way, with gamepieces stored in an array of gamepiece objects and the gameboard a simple character array that gets updated before rendering. there is no 'collision detection' yet, just sketched out a rough draft, it took a little bit to figure out how i could build a clean relationship between the gamepieces' data model and their on screen representation. the key was providing the gamepieces a reference to the world they live in, map[][], which makes total sense to me now, but took a rebuild to arrive at.