implemented basic collision detection
implemented bones of collision detection. check() is passed the index of an gamepiece in map, checks whether object is mobile, meaning it can be pushed around or not. I added this little flag with the future in mind, where it might be used if the player is on possession of some item that allows them to move certain things. in general this field makes it cleaner than having a large switch statement with a name field or checking the type of the gamepiece being encountered...on second thought though, what would be the difference? hmm...