game board stays centered on window resize
game board remiains centered in window when window is resized. key to getting it working was setMaximumSize on panel and using getMaximumSize to set the size of the image, it finally clicked when i realized that a dynamically set the drawing location, so it didn't matter how large the drawing surface was. took an hour of screwing around to figure it out.