rough implementation of random rooms
restructured the game world. game boards are now rooms in a dungeon. the size of the rooms is randomly determined (hail randu) when game class is initialized. player is placed in first room and game begins. next is to figure out how to connect the rooms together in a coherent manner.