THE DOORS WORK
so the doors work now. you can fully move through a dungeon. im not sure where the bugs are yet and im not 100 pcnt happy with the structure of the data changes that move the player through the dungeon, but it works and i can restructure later after ive implemented the rest of door placement.