rough game interface constructed
i added a panel to the frame to act as an info panel that will display player stats and list items. it will also take over displaying messages. ive made the window non-resizable for the time being as i have to figure out how to get the resize mechanism i built to play nice with the layout manager that places the panels side by side. im not familiar with layout managers just yet.